a0python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "a0python"
version = "0.1.0"
description = "a0 — PTCA-structured cognitive routing shell"
requires-python = ">=3.10"

dependencies = [
    "anyio>=4.0",
    "textual>=0.50",
    "gradio>=4.7",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
anthropic = [
    "anthropic>=0.25",
]
agent = [
    "claude-agent-sdk",
]
local-llama = [
    "llama-cpp-python>=0.2",
]
tools = [
    "pypdf>=4.0",
    "openai-whisper",
]
dev = [
    "pytest>=7.0",
]

[project.scripts]
a0 = "a0.a0:main"
a0-web = "a0.guardian.ui.web.app:main"

[tool.hatch.build.targets.wheel]
packages = ["a0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
